DRFapi/requirements.txt
asgiref==3.6.0
Django==4.2.1
django-cors-headers==4.0.0
django-environ==0.10.0
django-filter==1.1.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-filters==0.11.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
